# openapi-v2.yaml
openapi: 3.1.0
info:
  title: eLearning Template Engine API
  description: |
    Backend API for the composable component-based eLearning authoring platform.
    Supports 89+ component types, per-interaction audio, page completion tracking,
    weighted scoring, full theming (layout + color system), and SCORM 1.2/2004 export.
  version: 2.0.0
  contact:
    name: eLearning Platform Team
servers:
  - url: /api/v1
    description: API v1 namespace
...
paths:
...
  /courses/{courseId}/scoring:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      tags: [Scoring]
      operationId: getScoringConfig
      summary: Get scoring configuration
      responses:
        "200":
          description: Scoring configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoringConfigResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Scoring]
      operationId: updateScoringConfig
      summary: Update scoring configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoringConfigUpdateRequest"
      responses:
        "200":
          description: Updated scoring config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoringConfigResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /courses/{courseId}/scoring/validate:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    post:
      tags: [Scoring]
      operationId: validateScoringConfig
      summary: Validate scoring configuration
      description: Checks weights sum to 1.0, components are scorable, etc.
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResult"
        "404":
          $ref: "#/components/responses/NotFound"
  /courses/{courseId}/scoring/calculate:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    post:
      tags: [Scoring]
      operationId: calculateScore
      summary: Calculate score from submitted answers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoreCalculateRequest"
      responses:
        "200":
          description: Score calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreCalculateResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

  # ── Completion ──────────────────────────────
  /courses/{courseId}/completion:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      tags: [Completion]
      operationId: getCourseCompletion
      summary: Get course completion status
      responses:
        "200":
          description: Course completion breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseCompletionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /courses/{courseId}/pages/{pageId}/completion:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/PageId"
    get:
      tags: [Completion]
      operationId: getPageCompletion
      summary: Get page completion status
      responses:
        "200":
          description: Page completion breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageCompletionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Completion]
      operationId: recordPageCompletion
      summary: Record page completion event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageCompletionEvent"
      responses:
        "200":
          description: Recorded completion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageCompletionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /courses/{courseId}/interactions:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      tags: [Completion]
      operationId: listInteractions
      summary: List recorded interaction events
      parameters:
        - name: learnerId
          in: query
          schema:
            type: string
          description: Filter by learner ID
        - name: interactionType
          in: query
          schema:
            type: string
          description: Filter by interaction type
        - name: pageId
          in: query
          schema:
            type: string
          description: Filter by page ID
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of interaction events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InteractionEventResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Completion]
      operationId: recordInteraction
      summary: Record a component interaction event
      description: |
        Used in preview mode and non-SCORM deployments.
        For SCORM exports, tracking happens client-side via cmi.interactions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InteractionEvent"
      responses:
        "201":
          description: Interaction recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InteractionEventResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

  # ── Branching & Adaptive Navigation ─────
...
components:
...
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-zA-Z0-9_-]+$"
      description: Course identifier
    PageId:
      name: pageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Page UUID
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"

  # ── Schemas ─────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    ValidationErrorResponse:
      type: object
      required: [detail, errors]
      properties:
        detail:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    ValidationResult:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        warnings:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    ScormObjective:
      type: object
      properties:
        objectiveId:
          type: string
        componentIds:
          type: array
          items:
            type: string
          description: Components contributing to this objective
        description:
          type: string
        passingScore:
          type: number
          minimum: 0
          maximum: 100
    ScormReportingConfig:
      type: object
      properties:
        enabled:
          type: boolean
          default: true
        version:
          type: string
          enum: ["1.2", "2004"]
          default: "1.2"
        reportScore:
          type: boolean
          default: true
        reportCompletion:
          type: boolean
          default: true
        reportInteractions:
          type: boolean
          default: true
        objectives:
          type: array
          items:
            $ref: "#/components/schemas/ScormObjective"
    ComponentScoreConfig:
      type: object
      required: [componentId, maxPoints]
      properties:
        componentId:
          type: string
          format: uuid
        componentType:
          type: string
          description: Must be a scoringEnabled type
        weight:
          type: number
          minimum: 0
          maximum: 1
          description: Weight (sum must = 1.0 when weightedScoring=true)
        maxPoints:
          type: number
          minimum: 0
    ScoringConfig:
      type: object
      nullable: true
      properties:
        config:
          type: object
          properties:
            passingScore:
              type: number
              minimum: 0
              maximum: 100
              default: 70
            maxAttempts:
              type: integer
              nullable: true
              description: "null = unlimited"
            attemptScoring:
              type: string
              enum: [best, last, average]
              default: best
            showCorrectAnswers:
              type: boolean
              default: true
            showScoreAfterQuestion:
              type: boolean
              default: false
            showScoreAfterPage:
              type: boolean
              default: false
            weightedScoring:
              type: boolean
              default: false
            allowPartialCredit:
              type: boolean
              default: true
        componentScores:
          type: array
          items:
            $ref: "#/components/schemas/ComponentScoreConfig"
        scormReporting:
          $ref: "#/components/schemas/ScormReportingConfig"
    ScoringConfigResponse:
      allOf:
        - type: object
          properties:
            scoringId:
              type: string
              format: uuid
            courseId:
              type: string
        - $ref: "#/components/schemas/ScoringConfig"
    ScoringConfigUpdateRequest:
      $ref: "#/components/schemas/ScoringConfig"
    QuestionResponse:
      type: object
      properties:
        questionId:
          type: string
        selectedOptionIds:
          type: array
          items:
            type: string
          description: Selected option ID(s)
    ComponentAnswer:
      type: object
      required: [componentId, componentType, responses]
      properties:
        componentId:
          type: string
          format: uuid
        componentType:
          type: string
        responses:
          type: array
          items:
            $ref: "#/components/schemas/QuestionResponse"
    ScoreCalculateRequest:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/ComponentAnswer"
        attemptNumber:
          type: integer
          default: 1
    QuestionResult:
      type: object
      properties:
        questionId:
          type: string
        correct:
          type: boolean
        score:
          type: number
        maxScore:
          type: number
        partialCredit:
          type: boolean
          default: false
    ComponentResult:
      type: object
      properties:
        componentId:
          type: string
        componentType:
          type: string
        score:
          type: number
        maxScore:
          type: number
        weight:
          type: number
        weightedScore:
          type: number
        questionResults:
          type: array
          items:
            $ref: "#/components/schemas/QuestionResult"
    ScoreCalculateResponse:
      type: object
      properties:
        totalScore:
          type: number
        maxScore:
          type: number
        percentage:
          type: number
        passed:
          type: boolean
        passingScore:
          type: number
        componentResults:
          type: array
          items:
            $ref: "#/components/schemas/ComponentResult"
        attemptNumber:
          type: integer
        remainingAttempts:
          type: integer
          nullable: true

    # ── Completion ──────────────────────
    ComponentCompletionStatus:
      type: object
      properties:
        componentId:
          type: string
        completed:
          type: boolean
        completionType:
          type: string
          enum: [view, interact, audio, score, custom]
        threshold:
          type: number
          nullable: true
        audioProgress:
          type: object
          nullable: true
          properties:
            listened:
              type: integer
            required:
              type: integer
    PageCompletionResponse:
      type: object
      properties:
        pageId:
          type: string
        title:
          type: string
        completed:
          type: boolean
        strategy:
          type: string
          enum: [all, any, percentage, custom]
        components:
          type: array
          items:
            $ref: "#/components/schemas/ComponentCompletionStatus"
    CourseCompletionResponse:
      type: object
      properties:
        courseId:
          type: string
        status:
          type: string
          enum: [not-started, in-progress, completed, passed, failed]
        overallProgress:
          type: number
          description: Percentage 0-100
        pages:
          type: array
          items:
            $ref: "#/components/schemas/PageCompletionResponse"
    PageCompletionEvent:
      type: object
      required: [componentStates]
      properties:
        componentStates:
          type: array
          description: Component completion states for this page
          items:
            type: object
            required: [componentId, completed]
            properties:
              componentId:
                type: string
              completed:
                type: boolean
              interactionsCompleted:
                type: array
                items:
                  type: string
                description: Interaction IDs completed within the component
              audiosCompleted:
                type: array
                items:
                  type: string
                description: Audio IDs completed within the component
              score:
                type: number
                nullable: true

    # ── Interaction Events ──────────────
    InteractionData:
      type: object
      properties:
        interactionId:
          type: string
          nullable: true
          description: "E.g. 'tab-2'"
        value:
          description: Response value (selected option ID, text input, etc.)
        score:
          type: number
          nullable: true
        maxScore:
          type: number
          nullable: true
        isCorrect:
          type: boolean
          nullable: true
        duration:
          type: number
          nullable: true
          description: Seconds spent
    InteractionEvent:
      type: object
      required: [pageId, componentId, interactionType]
      properties:
        pageId:
          type: string
        componentId:
          type: string
        interactionType:
          type: string
          description: Open string (e.g. flip, reveal, drag-sort, text-input, rating, acknowledge, download)
        data:
          $ref: "#/components/schemas/InteractionData"
        completed:
          type: boolean
          default: false
    InteractionEventResponse:
      type: object
      properties:
        eventId:
          type: string
          format: uuid
        courseId:
          type: string
        pageId:
          type: string
        componentId:
          type: string
        learnerId:
          type: string
        interactionType:
          type: string
        data:
          $ref: "#/components/schemas/InteractionData"
        score:
          type: number
          nullable: true
        maxScore:
          type: number
          nullable: true
        duration:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        completed:
          type: boolean

    # ── Branching & Adaptive Navigation ─────
...